Renamed freediscovery.clustering->cluster
